Refactor to simplify directory structure
