--- mhcflurry/class1_affinity_prediction/__init__.py
@@ -1,9 +1,0 @@
-from __future__ import absolute_import
-
-from .class1_neural_network import Class1NeuralNetwork
-from .class1_affinity_predictor import Class1AffinityPredictor
-
-__all__ = [
-    'Class1NeuralNetwork',
-    'Class1AffinityPredictor',
-]
